rename useInterceptor to useNavigationPrompt
