+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.5.9] - Unpublished
+### Added
+- `useNavigationPrompt` for confirming navigation changes
+
 ## [0.5.8] - 2019-10-14
 ### Fixed
 - `<Link target="_blank" />` triggering local navigation
