+++ types/raviger.d.ts
@@ -64,4 +64,4 @@ export function useQueryParams<T, Q>(
   serializeFn?: (query: Q) => string
 ): [T, (query: Q, replace?: boolean) => void]
 
-export function useInterceptor(predicate = true, prompt?: string): void
+export function useNavigationPrompt(predicate = true, prompt?: string): void
